Fix the correctness defects found by the pre-release review
Five independent reviews ran over the tree before cutting the release. These
are the confirmed correctness defects from them, each verified against the
RocksDB 11.8.1 headers or by executing the failure. All five predate this
release.

Four event-listener enums did not match the native values. CompactionReason,
FlushReason and BackgroundErrorReason were shifted from the third member
onward and carried names for reasons RocksDB does not have, so a manual
compaction arrived as FilesMarkedForCompaction and an explicit flush as
WriteBufferManager. WriteStallCondition was worse: it was inverted. RocksDB
declares kDelayed, kStopped, kNormal in that order, because new conditions
are added before kNormal, so the onset of a write stall was reported as
Normal and recovery as Stopped. An application throttling on stalls acted on
the opposite signal.

Nothing caught this because the tests asserted the wrapper's own numbers.
They now check every member of all four enums against the header, and two
behavioural tests assert that a manual flush reports itself as a manual flush
and a manual compaction as a manual compaction. Both fail against the old
values.

A MergeOperator that did not override PartialMerge terminated the process on
flush. RocksDB calls the partial-merge slot through an unchecked function
pointer, unlike the delete-value slot immediately below it, and reaches it on
any flush that collapses two or more operands for one key. The wrapper
installed a null pointer whenever the method was not overridden. Confirmed by
running it: the test host died with an access violation. The slot is now
always installed, and the base implementation returns false, which is the
correct answer for an operator that cannot combine operands.

Three binding defects, none of them reachable from the wrapper yet:

- rocksdb_iter_key_slice and its two siblings return a 16-byte struct by
  value but were declared as returning a pointer. Windows x64 returns such a
  struct through a hidden pointer argument, so calling one would have written
  over the iterator and read the wrong register as its argument. The
  generator now emits the struct.
- Four functions returning C bool were declared as returning a pointer-sized
  integer, which reads register bits no ABI requires the callee to define.
  Seven bool parameters had the same declaration.
- Thirteen pointer-to-array parameters were untyped integers, including on
  rocksdb_open_column_families and rocksdb_multi_get_cf.

The root cause of all three, and of the four marshalling defects fixed
earlier in this release, is that the generator's type switch ended in two
arms that mapped anything unrecognised to a pointer-sized integer. It now
throws instead. Getting there meant giving explicit mappings to everything
that was reaching those arms: C bool, pointers to int, triple pointers, a
trailing const on a pointer, T* const*, the slice struct, and the nineteen
callback typedefs. Generation is byte-for-byte deterministic across runs.

+++ RocksDb.Net/RocksDb.cs
@@ -86,7 +86,7 @@ public sealed class RocksDb : RocksDbHandle
             fixed (nint* handlesPtr = cfHandles)
                 handle = NativeMethods.rocksdb_open_column_families(
                     options.Handle, path, count,
-                    namesPtr, (nint)optsPtr, handlesPtr, ref err);
+                    namesPtr, optsPtr, handlesPtr, ref err);
         }
         finally
         {
@@ -141,7 +141,7 @@ public sealed class RocksDb : RocksDbHandle
             fixed (nint* handlesPtr = cfHandles)
                 handle = NativeMethods.rocksdb_open_for_read_only_column_families(
                     options.Handle, path, count,
-                    namesPtr, (nint)optsPtr, handlesPtr,
+                    namesPtr, optsPtr, handlesPtr,
                     errorIfWalExists ? (byte)1 : (byte)0, ref err);
         }
         finally
